requirements.txt
# Python requirements for Plantopia DEM processing
gdal>=3.0.0
numpy>=1.19.0
Pillow>=8.0.0
scipy>=1.5.0
